JE2-77 [FE] Moved editing logic to ClientsDataList.
