+++ frontend/src/features/clients/containers/ClientDetail.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import {
   Box, Button,
@@ -13,22 +13,16 @@ import {
   EmailOutlined,
   PhoneAndroidOutlined,
   LocationOnOutlined,
-  AccountBalanceOutlined, EditOutlined, DeleteOutline,
+  AccountBalanceOutlined, DeleteOutline,
 } from '@mui/icons-material'
 import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts'
 import { selectClient, selectClientError, selectLoadingFetchClient } from '../../../store/slices/clientSlice.ts'
 import { fetchClientById } from '../../../store/thunks/clientThunk.ts'
 import ClientInfoItem from '../components/ClientInfoItem.tsx'
-import ClientForm from '../components/ClientForm.tsx'
-import Modal from '../../../components/UI/Modal/Modal.tsx'
 import { useClientsList } from '../hooks/useClientsList.ts'
 import { toast } from 'react-toastify'
 
 const ClientDetail = () => {
-  const [editModalOpen, setEditModalOpen] = useState(false)
-
-  const handleOpenEditModal = () => setEditModalOpen(true)
-  const handleCloseEditModal = () => setEditModalOpen(false)
 
   const { clientId } = useParams()
   const navigate = useNavigate()
@@ -141,18 +135,6 @@ const ClientDetail = () => {
             }}>
               <Button
                 variant="contained"
-                startIcon={<EditOutlined />}
-                onClick={handleOpenEditModal}
-                sx={{
-                  px: 3,
-                  borderRadius: 2,
-                  textTransform: 'none',
-                }}
-              >
-                Редактировать
-              </Button>
-              <Button
-                variant="contained"
                 color="error"
                 startIcon={<DeleteOutline />}
                 sx={{
@@ -176,9 +158,6 @@ const ClientDetail = () => {
               </Button>
             </Box>
           </Card>
-          <Modal open={editModalOpen} handleClose={handleCloseEditModal}>
-            <ClientForm client={client} onClose={handleCloseEditModal} />
-          </Modal>
         </Box>
       </>}
     </>
